put *_raw functions into a separate header.
