+++ clmctfy/include/clmctfy.h
@@ -21,18 +21,6 @@ struct container_api;
 //
 // Arguments:
 //  - s: [output] s will be used as output. It contains the error code/message.
-//  - spec: Serialized data (protobuf format) containing the specification.
-//  - spec_size: Size of the serialized data.
-//
-// Returns:
-//  Returns the error code. 0 on success. The return code is same as
-//  status_get_code(s).
-int lmctfy_init_machine_raw(struct status *s, const void *spec, const int spec_size);
-
-// Initializes the machine to start being able to create containers.
-//
-// Arguments:
-//  - s: [output] s will be used as output. It contains the error code/message.
 //  - spec: The specification.
 //
 // Returns:
@@ -74,13 +62,6 @@ int lmctfy_container_api_get_container(
     const struct container_api *api,
     const char *container_name);
 
-int lmctfy_container_api_create_container_raw(
-    struct status *s,
-    struct container **container,
-    struct container_api *api,
-    const char *container_name,
-    const void *spec,
-    const int spec_size);
 #ifdef __cplusplus
 }
 #endif // __cplusplus
